+++ solidblocks-cli/src/test/kotlin/de/solidblocks/cli/HetznerAsgTest.kt
@@ -1,6 +1,10 @@
 package de.solidblocks.cli
 
+import de.solidblocks.cli.hetzner.api.HetznerApi
 import de.solidblocks.cli.hetzner.asg.*
+import io.kotest.common.runBlocking
+import io.kotest.matchers.collections.shouldContain
+import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import java.util.*
@@ -11,6 +15,7 @@ import kotlin.time.ExperimentalTime
 class HetznerAsgTest {
 
     val hcloudToken = System.getenv("HCLOUD_TOKEN").toString()
+    val api = HetznerApi(hcloudToken)
 
     @OptIn(ExperimentalTime::class)
     fun createUserData() = """
@@ -22,7 +27,7 @@ class HetznerAsgTest {
 
     @Test
     fun testInvalidLoadBalancerName() {
-        HetznerAsg(hcloudToken).run(
+        HetznerAsg(hcloudToken).rotate(
             LoadBalancerReference("invalid"),
             LocationReference("nbg1"),
             ServerTypeReference("cx22"),
@@ -33,7 +38,7 @@ class HetznerAsgTest {
 
     @Test
     fun testInvalidLoadBalancerId() {
-        HetznerAsg(hcloudToken).run(
+        HetznerAsg(hcloudToken).rotate(
             LoadBalancerReference("0"),
             LocationReference("nbg1"),
             ServerTypeReference("cx22"),
@@ -44,50 +49,95 @@ class HetznerAsgTest {
 
     @Test
     fun testAsg1() {
-        HetznerAsg(hcloudToken).run(
-            LoadBalancerReference("application1"),
-            LocationReference("nbg1"),
-            ServerTypeReference("cx22"),
-            ImageReference("debian-12"),
-            userData = createUserData(),
-        ) shouldBe ASG_ROTATE_STATUS.OK
+        runBlocking {
 
-        HetznerAsg(hcloudToken).run(
-            LoadBalancerReference("application1"),
-            LocationReference("nbg1"),
-            ServerTypeReference("cx22"),
-            ImageReference("debian-12"),
-            userData = createUserData(),
-        ) shouldBe ASG_ROTATE_STATUS.OK
+            val unmanagedServer = api.servers.get("hcloud-server1")!!.server
+            var targets = api.loadBalancers.get("application1")!!.loadbalancer.targets
+            targets.map { it.server!!.id } shouldContain unmanagedServer.id
+            val initialServerCount = targets.count()
+
+            HetznerAsg(hcloudToken).rotate(
+                LoadBalancerReference("application1"),
+                LocationReference("nbg1"),
+                ServerTypeReference("cx22"),
+                ImageReference("debian-12"),
+                userData = createUserData(),
+            ) shouldBe ASG_ROTATE_STATUS.OK
+
+            targets = api.loadBalancers.get("application1")!!.loadbalancer.targets
+            targets.map { it.server!!.id } shouldContain unmanagedServer.id
+            targets shouldHaveSize initialServerCount
+
+            HetznerAsg(hcloudToken).rotate(
+                LoadBalancerReference("application1"),
+                LocationReference("nbg1"),
+                ServerTypeReference("cx22"),
+                ImageReference("debian-12"),
+                userData = createUserData(),
+                replicas = 2
+            ) shouldBe ASG_ROTATE_STATUS.OK
+
+            targets = api.loadBalancers.get("application1")!!.loadbalancer.targets
+            targets.map { it.server!!.id } shouldContain unmanagedServer.id
+            targets shouldHaveSize initialServerCount + 1
+
+            val userData = createUserData()
+
+            HetznerAsg(hcloudToken).rotate(
+                LoadBalancerReference("application1"),
+                LocationReference("nbg1"),
+                ServerTypeReference("cx22"),
+                ImageReference("debian-12"),
+                userData = userData,
+                replicas = 1
+            ) shouldBe ASG_ROTATE_STATUS.OK
+
+            targets = api.loadBalancers.get("application1")!!.loadbalancer.targets
+            targets.map { it.server!!.id } shouldContain unmanagedServer.id
+            targets shouldHaveSize initialServerCount
+
+            HetznerAsg(hcloudToken).rotate(
+                LoadBalancerReference("application1"),
+                LocationReference("nbg1"),
+                ServerTypeReference("cx22"),
+                ImageReference("debian-12"),
+                userData = userData,
+                replicas = 1
+            ) shouldBe ASG_ROTATE_STATUS.OK
+
+            targets = api.loadBalancers.get("application1")!!.loadbalancer.targets
+            targets.map { it.server!!.id } shouldContain unmanagedServer.id
+            targets shouldHaveSize initialServerCount
+        }
     }
 
     @Test
     fun testAsg1Timeout() {
-        HetznerAsg(hcloudToken).run(
+        HetznerAsg(hcloudToken).rotate(
             LoadBalancerReference("application1"),
             LocationReference("nbg1"),
             ServerTypeReference("cx22"),
             ImageReference("debian-12"),
             userData = "invalid userdata ${UUID.randomUUID()}",
-            healthTimeout = 45.seconds,
-            totalTimeout = 60.seconds
+            healthTimeout = 60.seconds,
+            totalTimeout = 75.seconds
         ) shouldBe ASG_ROTATE_STATUS.TIMEOUT
     }
 
     @Test
     fun testAsg2() {
-        HetznerAsg(hcloudToken).run(
+        HetznerAsg(hcloudToken).rotate(
             LoadBalancerReference("application2"),
             LocationReference("nbg1"),
             ServerTypeReference("cx22"),
             ImageReference("debian-12"),
-            userData = "invalid userdata ${UUID.randomUUID()}",
-        ) shouldBe ASG_ROTATE_STATUS.TIMEOUT
+            userData = createUserData(),
+        ) shouldBe ASG_ROTATE_STATUS.OK
     }
 
     @Test
     fun testAsg3() {
-        HetznerAsg(hcloudToken).run(
+        HetznerAsg(hcloudToken).rotate(
             LoadBalancerReference("application3"),
             LocationReference("nbg1"),
             ServerTypeReference("cx22"),